+++ pkg/server/errors/errors_test.go
@@ -68,9 +68,9 @@ func TestHandleErrors(t *testing.T) {
 			storageErr:              context.DeadlineExceeded,
 			expectedTranslatedError: ErrRequestDeadlineExceeded,
 		},
-		`passthrough_error`: {
-			storageErr:              NewPassthroughError(NewEncodedError(1, "oh no")),
-			expectedTranslatedError: ErrPassthrough,
+		`throttling_error`: {
+			storageErr:              storage.ErrTransactionThrottled,
+			expectedTranslatedError: ErrTransactionThrottled,
 		},
 	}
 	for testName, test := range tests {